Move measurements to their own file
