+++ lib/Common/utils.h
@@ -3,8 +3,6 @@
 #define GAGGIA_UTILS_H
 
 #include "Arduino.h"
-#include <deque>
-#include <type_traits>
 
 enum class TransitionCurve {
   EASE_IN_OUT,
@@ -16,35 +14,4 @@ enum class TransitionCurve {
 
 float mapRange(float sourceNumber, float fromA, float fromB, float toA, float toB, int decimalPrecision, TransitionCurve transition = TransitionCurve::LINEAR);
 
-struct Measurement {
-  float value;
-  uint32_t millis;
-};
-
-struct MeasurementChange {
-  float deltaValue;
-  uint32_t deltaMillis;
-
-  float changeSpeed() {
-    if (deltaMillis == 0) {
-      return 0.f;
-    }
-    return 1000.f * deltaValue / static_cast<float>(deltaMillis);
-  }
-};
-
-class Measurements {
-private:
-  size_t size;
-  std::deque<Measurement> values;
-public:
-  Measurements(size_t size) : size(size) {}
-
-  Measurement latest();
-  Measurement previous();
-  MeasurementChange measurementChange();
-  void clear();
-  void add(float value);
-  void add(Measurement measurement);
-};
 #endif
