+++ src/peripherals/remote_scales.h
@@ -1,7 +1,7 @@
 #ifndef REMOTE_SCALES_H
 #define REMOTE_SCALES_H
 
-#include "utils.h"
+#include "measurements.h"
 
 void remoteScalesTare(void);
 Measurement remoteScalesGetWeight(void);
